+++ cloudant-client/src/test/resources/design-files/query_design_doc.js
@@ -1,0 +1,17 @@
+{
+ "_id": "_design/testQuery",
+ "language": "query",
+ "views": {
+  "testView": {
+   "map": {"fields": {"Person_dob": "asc"}},
+   "reduce": "_count",
+   "options": {
+    "def": {
+     "fields": [
+      "Person_dob"
+     ]
+    }
+   }
+  }
+ }
+}